# webAdmin/openapi.yaml
openapi: 3.1.0
info:
  title: Roulette Reward Backend API
  version: 1.1.0
  description: |
    실제 컨트롤러 구현 기준 API 명세입니다.
    인증은 JWT가 아닌 커스텀 헤더(`X-USER-ID`, `X-ADMIN-TOKEN`)를 사용합니다.
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      operationId: mockLogin
      summary: 모킹 로그인
      description: 닉네임으로 사용자 식별자를 발급/조회합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MockLoginRequest'
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseMockLogin'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/auth/admin-login:
    post:
      tags: [Authentication]
      operationId: adminLogin
      summary: 어드민 인증
      description: 특정 인증 코드 입력 시 고정 어드민 토큰을 반환합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        '200':
          description: 인증 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAdminLogin'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/admin/roulette/participants/count:
    get:
      tags: [AdminRoulette]
      operationId: getRouletteParticipantCount
      summary: 참여자 수 조회(어드민)
      parameters:
        - $ref: '#/components/parameters/XAdminTokenHeader'
        - $ref: '#/components/parameters/DateQuery'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRouletteParticipantCount'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/roulette/participants:
    get:
      tags: [AdminRoulette]
      operationId: getRouletteParticipants
      summary: 참여자 목록 조회(어드민)
      parameters:
        - $ref: '#/components/parameters/XAdminTokenHeader'
        - $ref: '#/components/parameters/DateQuery'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRouletteParticipantList'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/roulette/participations/{participationId}/cancel:
    post:
      tags: [AdminRoulette]
      operationId: cancelRouletteParticipation
      summary: 룰렛 참여 취소(어드민)
      parameters:
        - $ref: '#/components/parameters/XAdminTokenHeader'
        - $ref: '#/components/parameters/ParticipationIdPath'
      responses:
        '200':
          description: 취소 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseCancelRouletteParticipation'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
...
    XAdminTokenHeader:
      name: X-ADMIN-TOKEN
      in: header
      required: true
      description: 어드민 인증 토큰
      schema:
        type: string
      example: ADMIN-STATIC-TOKEN
...
    DateQuery:
      name: date
      in: query
      required: false
      schema:
        type: string
        format: date
      example: '2026-02-15'
    ParticipationIdPath:
      name: participationId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 9001
...
  responses:
    BadRequestError:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    UnauthorizedError:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ForbiddenError:
      description: 권한 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFoundError:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ConflictError:
      description: 상태 충돌
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    InternalServerError:
      description: 내부 서버 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
    ApiErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
    ErrorBody:
      type: object
      required: [code, message, details, trace_id]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
        trace_id:
          type: string
    ErrorDetail:
      type: object
      required: [field, reason]
      properties:
        field:
          type: string
        reason:
          type: string
    UserRole:
      type: string
      enum: [USER, ADMIN]
...
    MockLoginRequest:
      type: object
      required: [nickname]
      properties:
        nickname:
          type: string
          minLength: 2
          maxLength: 30
          pattern: '^[a-zA-Z0-9가-힣_-]+$'
    AdminLoginRequest:
      type: object
      required: [adminCode]
      properties:
        adminCode:
          type: string
          minLength: 1
    MockLoginResponse:
      type: object
      required: [userId, nickname, role]
      properties:
        userId:
          type: integer
          format: int64
        nickname:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
    AdminLoginResponse:
      type: object
      required: [headerName, adminToken]
      properties:
        headerName:
          type: string
          example: X-ADMIN-TOKEN
        adminToken:
          type: string
          example: ADMIN-STATIC-TOKEN
...
    RouletteParticipantCountResponse:
      type: object
      required: [date, participantCount, totalAwardedPoints]
      properties:
        date:
          type: string
          format: date
        participantCount:
          type: integer
          format: int64
        totalAwardedPoints:
          type: integer
    RouletteParticipantItemResponse:
      type: object
      required: [participationId, userId, awardedPoints, awardedAt, canceled]
      properties:
        participationId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        awardedPoints:
          type: integer
        awardedAt:
          type: string
          format: date-time
        canceled:
          type: boolean
    RouletteParticipantListResponse:
      type: object
      required: [date, totalItems, items]
      properties:
        date:
          type: string
          format: date
        totalItems:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/RouletteParticipantItemResponse'
    CancelRouletteParticipationResponse:
      type: object
      required: [participationId, userId, recoveredPoints, canceledAt]
      properties:
        participationId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        recoveredPoints:
          type: integer
        canceledAt:
          type: string
          format: date-time
...
    ApiResponseMockLogin:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/MockLoginResponse'
    ApiResponseAdminLogin:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/AdminLoginResponse'
...
    ApiResponseRouletteParticipantCount:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/RouletteParticipantCountResponse'
    ApiResponseRouletteParticipantList:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/RouletteParticipantListResponse'
    ApiResponseCancelRouletteParticipation:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/CancelRouletteParticipationResponse'
...